backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-filter==21.1
six==1.16.0
django-import-export==2.6.1
## The following requirements were added by pip freeze:
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
et-xmlfile==1.1.0
idna==3.3
jwcrypto==1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.7
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
django-oauth-toolkit==1.5.0
drf-social-oauth2==1.0.9
oauthlib==3.1.1
psycopg2==2.9.3
requests-oauthlib==1.3.0
boto3==1.24.57
botocore==1.27.57
django-storages==1.13.1
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
django-multiselectfield==0.1.12
